+++ core/commands/swarm.go
@@ -7,25 +7,22 @@ import (
 	"errors"
 	"fmt"
 	"io"
-	"os"
 	"path"
 	"sort"
-	"strings"
 	"sync"
 	"time"
 
+	files "github.com/ipfs/go-ipfs-files"
 	"github.com/ipfs/go-ipfs/commands"
 	"github.com/ipfs/go-ipfs/config"
 	"github.com/ipfs/go-ipfs/core/commands/cmdenv"
+	"github.com/ipfs/go-ipfs/core/node/libp2p"
 	"github.com/ipfs/go-ipfs/repo"
 	"github.com/ipfs/go-ipfs/repo/fsrepo"
 
 	cmds "github.com/ipfs/go-ipfs-cmds"
-	"github.com/libp2p/go-libp2p-core/network"
 	inet "github.com/libp2p/go-libp2p-core/network"
 	"github.com/libp2p/go-libp2p-core/peer"
-	"github.com/libp2p/go-libp2p-core/protocol"
-	rcmgr "github.com/libp2p/go-libp2p-resource-manager"
 	ma "github.com/multiformats/go-multiaddr"
 	madns "github.com/multiformats/go-multiaddr-dns"
 	mamask "github.com/whyrusleeping/multiaddr-filter"
@@ -59,8 +56,8 @@ ipfs peers in the internet.
 		"filters":    swarmFiltersCmd,
 		"peers":      swarmPeersCmd,
 		"peering":    swarmPeeringCmd,
-		"stats":      swarmStatsCmd,
-		"limit":      swarmLimitCmd,
+		"stats":      swarmStatsCmd, // libp2p Network Resource Manager
+		"limit":      swarmLimitCmd, // libp2p Network Resource Manager
 	},
 }
 
@@ -318,13 +315,15 @@ var swarmStatsCmd = &cmds.Command{
 	Helptext: cmds.HelpText{
 		Tagline: "Report resource usage for a scope.",
 		LongDescription: `Report resource usage for a scope.
-  The scope can be one of the following:
-  - system        -- reports the system aggregate resource usage.
-  - transient     -- reports the transient resource usage.
-  - svc:<service> -- reports the resource usage of a specific service.
-  - proto:<proto> -- reports the resource usage of a specific protocol.
-  - peer:<peer>   -- reports the resource usage of a specific peer.
-  - all           -- reports the resource usage for all currently active scopes.
+The scope can be one of the following:
+- system        -- reports the system aggregate resource usage.
+- transient     -- reports the transient resource usage.
+- svc:<service> -- reports the resource usage of a specific service.
+- proto:<proto> -- reports the resource usage of a specific protocol.
+- peer:<peer>   -- reports the resource usage of a specific peer.
+- all           -- reports the resource usage for all currently active scopes.
+
+The output of this command is JSON.
 `},
 	Arguments: []cmds.Argument{
 		cmds.StringArg("scope", true, false, "scope of the stat report"),
@@ -343,7 +342,7 @@ var swarmStatsCmd = &cmds.Command{
 			return fmt.Errorf("must specify exactly one scope")
 		}
 		scope := req.Arguments[0]
-		result, err := NetStat(node.ResourceManager, req.Context, scope)
+		result, err := libp2p.NetStat(node.ResourceManager, scope)
 		if err != nil {
 			return err
 		}
@@ -356,6 +355,9 @@ var swarmStatsCmd = &cmds.Command{
 		}
 		return cmds.EmitOnce(res, b)
 	},
+	Encoders: cmds.EncoderMap{
+		cmds.Text: HumanJsonEncoder,
+	},
 }
 
 var swarmLimitCmd = &cmds.Command{
@@ -363,20 +365,27 @@ var swarmLimitCmd = &cmds.Command{
 	Helptext: cmds.HelpText{
 		Tagline: "Get or set resource limits for a scope.",
 		LongDescription: `Get or set resource limits for a scope.
-  The scope can be one of the following:
-  - system        -- reports the system aggregate resource usage.
-  - transient     -- reports the transient resource usage.
-  - svc:<service> -- reports the resource usage of a specific service.
-  - proto:<proto> -- reports the resource usage of a specific protocol.
-  - peer:<peer>   -- reports the resource usage of a specific peer.
- The limit is json-formatted, with the same structure as the limits file.
+The scope can be one of the following:
+- system        -- limits for the system aggregate resource usage.
+- transient     -- limits for the transient resource usage.
+- svc:<service> -- limits for the resource usage of a specific service.
+- proto:<proto> -- limits for the resource usage of a specific protocol.
+- peer:<peer>   -- limits for the resource usage of a specific peer.
+
+The output of this command is JSON.
+
+It is possible to use this command to inspect and tweak limits at runtime:
+
+	$ ipfs swarm limit system > limit.json
+	$ vi limit.json
+	$ ipfs swarm limit system limit.json
+
+Changes made via command line are discarded on node shutdown.
+For permanent limits set Swarm.ResourceMgr.Limits in the $IPFS_PATH/config file.
 `},
 	Arguments: []cmds.Argument{
 		cmds.StringArg("scope", true, false, "scope of the limit"),
-		cmds.StringArg("limit", false, false, "path of the limit configuration file"),
-	},
-	Options: []cmds.Option{
-		cmds.BoolOption("set", "s", "Set the limit for a scope (instead of viewing it).").WithDefault(false),
+		cmds.FileArg("limit.json", false, false, "limits to be set").EnableStdin(),
 	},
 	Run: func(req *cmds.Request, res cmds.ResponseEmitter, env cmds.Environment) error {
 		node, err := cmdenv.GetNode(env)
@@ -388,33 +397,29 @@ var swarmLimitCmd = &cmds.Command{
 			return fmt.Errorf("no resource manager available, make sure the daemon is running")
 		}
 
-		setLimit, _ := req.Options["set"].(bool)
-		if setLimit {
-			if len(req.Arguments) != 2 {
-				return fmt.Errorf("must specify exactly a scope and a limit")
-			}
+		scope := req.Arguments[0]
 
-			scope := req.Arguments[0]
-			limitPath := req.Arguments[1]
-			limitStr, err := os.ReadFile(limitPath)
-			if err != nil {
+		//  set scope limit to new values (when limit.json is passed as a second arg)
+		if req.Files != nil {
+			var newLimit config.ResourceMgrScopeConfig
+			it := req.Files.Entries()
+			if it.Next() {
+				file := files.FileFromEntry(it)
+				if file == nil {
+					return errors.New("expected a JSON file")
+				}
+				if err := json.NewDecoder(file).Decode(&newLimit); err != nil {
+					return errors.New("failed to decode JSON as ResourceMgrScopeConfig")
+				}
+				return libp2p.NetSetLimit(node.ResourceManager, scope, newLimit)
+			}
+			if err := it.Err(); err != nil {
 				return fmt.Errorf("error opening limit JSON file: %w", err)
 			}
-
-			var limit NetLimitConfig
-			err = json.Unmarshal([]byte(limitStr), &limit)
-			if err != nil {
-				return fmt.Errorf("error decoding limit: %w", err)
-			}
-
-			return NetSetLimit(node.ResourceManager, req.Context, scope, limit)
 		}
 
-		if len(req.Arguments) != 1 {
-			return fmt.Errorf("must specify exactly one scope")
-		}
-		scope := req.Arguments[0]
-		result, err := NetLimit(node.ResourceManager, req.Context, scope)
+		// get scope limit
+		result, err := libp2p.NetLimit(node.ResourceManager, scope)
 		if err != nil {
 			return err
 		}
@@ -427,289 +432,9 @@ var swarmLimitCmd = &cmds.Command{
 		}
 		return cmds.EmitOnce(res, b)
 	},
-}
-
-// FIXME(BLOCKING): Decide where to move the net stat/limit logic and types.
-
-type NetStatOut struct {
-	System    *network.ScopeStat           `json:",omitempty"`
-	Transient *network.ScopeStat           `json:",omitempty"`
-	Services  map[string]network.ScopeStat `json:",omitempty"`
-	Protocols map[string]network.ScopeStat `json:",omitempty"`
-	Peers     map[string]network.ScopeStat `json:",omitempty"`
-}
-
-type NetLimitConfig struct {
-	Dynamic bool `json:",omitempty"`
-	// set if Dynamic is false
-	Memory int64 `json:",omitempty"`
-	// set if Dynamic is true
-	MemoryFraction float64 `json:",omitempty"`
-	MinMemory      int64   `json:",omitempty"`
-	MaxMemory      int64   `json:",omitempty"`
-
-	Streams, StreamsInbound, StreamsOutbound int
-	Conns, ConnsInbound, ConnsOutbound       int
-	FD                                       int
-}
-
-func NetStat(mgr network.ResourceManager, ctx context.Context, scope string) (NetStatOut, error) {
-	var err error
-	var result NetStatOut
-	switch {
-	case scope == "all":
-		rapi, ok := mgr.(rcmgr.ResourceManagerState)
-		if !ok {
-			return result, fmt.Errorf("resource manager does not support ResourceManagerState API")
-		}
-
-		stat := rapi.Stat()
-		result.System = &stat.System
-		result.Transient = &stat.Transient
-		if len(stat.Services) > 0 {
-			result.Services = stat.Services
-		}
-		if len(stat.Protocols) > 0 {
-			result.Protocols = make(map[string]network.ScopeStat, len(stat.Protocols))
-			for proto, stat := range stat.Protocols {
-				result.Protocols[string(proto)] = stat
-			}
-		}
-		if len(stat.Peers) > 0 {
-			result.Peers = make(map[string]network.ScopeStat, len(stat.Peers))
-			for p, stat := range stat.Peers {
-				result.Peers[p.Pretty()] = stat
-			}
-		}
-
-		return result, nil
-
-	case scope == "system":
-		err = mgr.ViewSystem(func(s network.ResourceScope) error {
-			stat := s.Stat()
-			result.System = &stat
-			return nil
-		})
-		return result, err
-
-	case scope == "transient":
-		err = mgr.ViewTransient(func(s network.ResourceScope) error {
-			stat := s.Stat()
-			result.Transient = &stat
-			return nil
-		})
-		return result, err
-
-	case strings.HasPrefix(scope, "svc:"):
-		svc := scope[4:]
-		err = mgr.ViewService(svc, func(s network.ServiceScope) error {
-			stat := s.Stat()
-			result.Services = map[string]network.ScopeStat{
-				svc: stat,
-			}
-			return nil
-		})
-		return result, err
-
-	case strings.HasPrefix(scope, "proto:"):
-		proto := scope[6:]
-		err = mgr.ViewProtocol(protocol.ID(proto), func(s network.ProtocolScope) error {
-			stat := s.Stat()
-			result.Protocols = map[string]network.ScopeStat{
-				proto: stat,
-			}
-			return nil
-		})
-		return result, err
-
-	case strings.HasPrefix(scope, "peer:"):
-		p := scope[5:]
-		pid, err := peer.Decode(p)
-		if err != nil {
-			return result, fmt.Errorf("invalid peer ID: %s: %w", p, err)
-		}
-		err = mgr.ViewPeer(pid, func(s network.PeerScope) error {
-			stat := s.Stat()
-			result.Peers = map[string]network.ScopeStat{
-				p: stat,
-			}
-			return nil
-		})
-		return result, err
-
-	default:
-		return result, fmt.Errorf("invalid scope %s", scope)
-	}
-}
-
-func NetLimit(mgr network.ResourceManager, ctx context.Context, scope string) (NetLimitConfig, error) {
-	var result NetLimitConfig
-	getLimit := func(s network.ResourceScope) error {
-		limiter, ok := s.(rcmgr.ResourceScopeLimiter)
-		if !ok {
-			return fmt.Errorf("resource scope doesn't implement ResourceScopeLimiter interface")
-		}
-
-		limit := limiter.Limit()
-		switch l := limit.(type) {
-		case *rcmgr.StaticLimit:
-			result.Memory = l.Memory
-			result.Streams = l.BaseLimit.Streams
-			result.StreamsInbound = l.BaseLimit.StreamsInbound
-			result.StreamsOutbound = l.BaseLimit.StreamsOutbound
-			result.Conns = l.BaseLimit.Conns
-			result.ConnsInbound = l.BaseLimit.ConnsInbound
-			result.ConnsOutbound = l.BaseLimit.ConnsOutbound
-			result.FD = l.BaseLimit.FD
-
-		case *rcmgr.DynamicLimit:
-			result.Dynamic = true
-			result.MemoryFraction = l.MemoryLimit.MemoryFraction
-			result.MinMemory = l.MemoryLimit.MinMemory
-			result.MaxMemory = l.MemoryLimit.MaxMemory
-			result.Streams = l.BaseLimit.Streams
-			result.StreamsInbound = l.BaseLimit.StreamsInbound
-			result.StreamsOutbound = l.BaseLimit.StreamsOutbound
-			result.Conns = l.BaseLimit.Conns
-			result.ConnsInbound = l.BaseLimit.ConnsInbound
-			result.ConnsOutbound = l.BaseLimit.ConnsOutbound
-			result.FD = l.BaseLimit.FD
-
-		default:
-			return fmt.Errorf("unknown limit type %T", limit)
-		}
-
-		return nil
-	}
-
-	switch {
-	case scope == "system":
-		err := mgr.ViewSystem(func(s network.ResourceScope) error {
-			return getLimit(s)
-		})
-		return result, err
-
-	case scope == "transient":
-		err := mgr.ViewTransient(func(s network.ResourceScope) error {
-			return getLimit(s)
-		})
-		return result, err
-
-	case strings.HasPrefix(scope, "svc:"):
-		svc := scope[4:]
-		err := mgr.ViewService(svc, func(s network.ServiceScope) error {
-			return getLimit(s)
-		})
-		return result, err
-
-	case strings.HasPrefix(scope, "proto:"):
-		proto := scope[6:]
-		err := mgr.ViewProtocol(protocol.ID(proto), func(s network.ProtocolScope) error {
-			return getLimit(s)
-		})
-		return result, err
-
-	case strings.HasPrefix(scope, "peer:"):
-		p := scope[5:]
-		pid, err := peer.Decode(p)
-		if err != nil {
-			return result, fmt.Errorf("invalid peer ID: %s: %w", p, err)
-		}
-		err = mgr.ViewPeer(pid, func(s network.PeerScope) error {
-			return getLimit(s)
-		})
-		return result, err
-
-	default:
-		return result, fmt.Errorf("invalid scope %s", scope)
-	}
-}
-
-func NetSetLimit(mgr network.ResourceManager, ctx context.Context, scope string, limit NetLimitConfig) error {
-	setLimit := func(s network.ResourceScope) error {
-		limiter, ok := s.(rcmgr.ResourceScopeLimiter)
-		if !ok {
-			return fmt.Errorf("resource scope doesn't implement ResourceScopeLimiter interface")
-		}
-
-		var newLimit rcmgr.Limit
-		if limit.Dynamic {
-			newLimit = &rcmgr.DynamicLimit{
-				MemoryLimit: rcmgr.MemoryLimit{
-					MemoryFraction: limit.MemoryFraction,
-					MinMemory:      limit.MinMemory,
-					MaxMemory:      limit.MaxMemory,
-				},
-				BaseLimit: rcmgr.BaseLimit{
-					Streams:         limit.Streams,
-					StreamsInbound:  limit.StreamsInbound,
-					StreamsOutbound: limit.StreamsOutbound,
-					Conns:           limit.Conns,
-					ConnsInbound:    limit.ConnsInbound,
-					ConnsOutbound:   limit.ConnsOutbound,
-					FD:              limit.FD,
-				},
-			}
-		} else {
-			newLimit = &rcmgr.StaticLimit{
-				Memory: limit.Memory,
-				BaseLimit: rcmgr.BaseLimit{
-					Streams:         limit.Streams,
-					StreamsInbound:  limit.StreamsInbound,
-					StreamsOutbound: limit.StreamsOutbound,
-					Conns:           limit.Conns,
-					ConnsInbound:    limit.ConnsInbound,
-					ConnsOutbound:   limit.ConnsOutbound,
-					FD:              limit.FD,
-				},
-			}
-		}
-
-		limiter.SetLimit(newLimit)
-		return nil
-	}
-
-	switch {
-	case scope == "system":
-		err := mgr.ViewSystem(func(s network.ResourceScope) error {
-			return setLimit(s)
-		})
-		return err
-
-	case scope == "transient":
-		err := mgr.ViewTransient(func(s network.ResourceScope) error {
-			return setLimit(s)
-		})
-		return err
-
-	case strings.HasPrefix(scope, "svc:"):
-		svc := scope[4:]
-		err := mgr.ViewService(svc, func(s network.ServiceScope) error {
-			return setLimit(s)
-		})
-		return err
-
-	case strings.HasPrefix(scope, "proto:"):
-		proto := scope[6:]
-		err := mgr.ViewProtocol(protocol.ID(proto), func(s network.ProtocolScope) error {
-			return setLimit(s)
-		})
-		return err
-
-	case strings.HasPrefix(scope, "peer:"):
-		p := scope[5:]
-		pid, err := peer.Decode(p)
-		if err != nil {
-			return fmt.Errorf("invalid peer ID: %s: %w", p, err)
-		}
-		err = mgr.ViewPeer(pid, func(s network.PeerScope) error {
-			return setLimit(s)
-		})
-		return err
-
-	default:
-		return fmt.Errorf("invalid scope %s", scope)
-	}
+	Encoders: cmds.EncoderMap{
+		cmds.Text: HumanJsonEncoder,
+	},
 }
 
 type streamInfo struct {
