+++ core/node/groups.go
@@ -145,7 +145,7 @@ func LibP2P(bcfg *BuildCfg, cfg *config.Config) fx.Option {
 		BaseLibP2P,
 
 		// Services (resource management)
-		fx.Provide(libp2p.ResourceManager()),
+		fx.Provide(libp2p.ResourceManager(cfg.Swarm.ResourceMgr)),
 
 		fx.Provide(libp2p.AddrFilters(cfg.Swarm.AddrFilters)),
 		fx.Provide(libp2p.AddrsFactory(cfg.Addresses.Announce, cfg.Addresses.AppendAnnounce, cfg.Addresses.NoAnnounce)),
